Frogger 0.9
+ Added sound effects and game music to the game

+++ Assets/Scripts/Settings.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public static class Settings
 {
-    public static int MusicVolume;
-    public static int EffectsVolume;
+    public static float MusicVolume;
+    public static float EffectsVolume;
     public static int Difficulty;
 
     public static int ScoreCount;
